backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
ultralytics>=8.1.0
Pillow==10.0.1
opencv-python-headless==4.8.0.76
